// apps/frontend/client/src/app/features/projects/enums/project-role.enum.ts
/**
 * PROJECTS ROLE ENUM
 * ---------------------------------
 * Author: Cleopatra Kwenda
 * Date: 2026-07-17
 * Purpose: will store all the users roles for the projects in one place.
 * Related Requirement: N/A
 */

export enum ProjectRole {
  DEVELOPER = 'Developer',
  MANAGER = 'Manager',
  ADMIN = 'Admin',
}