// apps/frontend/client/src/app/features/projects/enums/project-status.enum.ts
/**
 * PROJECTS STATUS ENUM
 * ---------------------------------
 * Author: Cleopatra Kwenda
 * Date: 2026-07-17
 * Purpose: to store all our different project statuses in one place.
 * Related Requirement: N/A
 * Responsibilities:
 *  -will display project stats
 */

export enum ProjectStatus {
  ACTIVE = 'Active',
  ON_HOLD = 'On Hold',
  COMPLETED = 'Completed',
  ARCHIVED = 'Archived',
}